up api voucher (CRU)

+++ CDWeb/src/main/java/org/example/cdweb_be/entity/Voucher.java
@@ -28,7 +28,7 @@ public class Voucher {
     String imagePath;
     String description;
     int quantity;
-    double percentDecrease;
+    int percentDecrease;
     int maxDecrease;
     int minPrice;
     int type;
